requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
torch
torchvision
pillow
opencv-python
tqdm
